Use key in dict test directly, iterate over items
The more usual way of testing for a keys existence in a dictionary is to use `key in dict` directly, not testing for membership in the key view returned by `.keys()`. There was also a loop that can be made simpler by using the `.items()` view to iterate over `key,value` pairs instead of needing to do a lookup.

Also didn't seem like deleting the old list of players was really necessary, the only thing that would do is trigger a KeyError if a previous iteration had somehow managed to load a json file without the players key (even if there was somehow a game with no players the c++ program should still write the key in the object with a null value).